Create a file to store data
The list is to be downloaded once a day to build history of prices and
generally to collect data about desired product or set of leggins. Each
list is programmed to be stored in a seperate file. Thanks to a content
of the file some kind of comparisons and predictions related to price
changes will be possible to make.